Refactored the Parking Space Managers
- They are now in separate files
- They have their own onSpaceClaimed methods cos the shared one was a bit broken.
