+++ ProjectFiles/Assets/Scripts/Game/Core/Parking/ClientParkingSpaceManager.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 3526576677914ecb9501e82df953c356
+timeCreated: 1583431449
