+++ ProjectFiles/Assets/Scripts/Game/Core/Parking/ServerParkingSpaceManager.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: 554f9354f10642b4aad238f1cd832486
+timeCreated: 1583431437
